Deal with issue when moves change direction
There's an issue with the iterative planner which goes back to Marlin or earlier.  Namely, the iterations (1) strive to increase the segment endpoint speeds, (2) make the endpoint speeds match at the junction between two segments, and (3) the code only sees the magnitude of the velocity vector and doesn't know the direction.  Net, net an end speed of -5mm/s and following starting speed of +15mm/s may satisfy the max speed change requirement of 20mm/s.  But then the iterative process, which is blind to direction, will try to lift the endpoint speed and end up making it -15 mm/s or even a little more.  Suddenly, the speed change is now 30 mm/s and not 20 mm/s.
We try to deal with this by pinching down the starting speed of the current segment to 0.  Calculate_trapezoid_for_block() will actually lift that a little with its minimal step rate check.