re-calculate homing timeout in Command
for a bigger bot, sometimes we have very long axis, which would take
more time to home.

since there's no convinient way to set homing timeout in MakerBot
Desktop and ReplicatorG, we just ignore the homing timeout in the
command completely, and re-calculate the optimal timeout according
to axis length and feedrate.