Change acceleration limit to be half of buffer size
This seems more reasonable. I don't see why this should slow down a print for
anybody that didn't need it slowed down already, but since the slowdown is
graceful, in the default case starting at 7/8ths speed is unlikely to be
noticed by anyone, except for extremely detailed prints that really needed it.