Pipeline sdcard reads
Sdcard reads are very inefficient, although we fill the command buffer, we
only read a single move command at a time. When printing with high detail
this is often not fast enough to keep the move command ring full, which
results in planner badness in addition to slowing down the print. So, if we
encounter a stream of move commands, pipeline those reads into the planner.
Additionally, if we still haven't filled the buffer blocks, read more from
the sdcard, and continue doing so until we do. This effectively prioritizes
reading of sdcard data during prints over everything else, except interrupts
of course. If a very long area of high detail is encountered, we can be in
that loop for a long time, since st_interrupt basically 100% cpu and is
draining the block buffer continuously. Therefore we break out after we've
done a bunch of work so that the menu and other slices are still responsive.